add: asdieButton,chooseMode
