+++ app.js
@@ -4,8 +4,8 @@ import { clickNewsStand } from "./js/newsStand.js";
 import { renderMain } from "./js/renderMain.js";
 
 (function init() {
-  clickNewsStand();
+  clickNewsStand(document.getElementById("header-div-01"));
   setDate(document.getElementById("header-div-02"));
   rollingBanner();
-  renderMain(true);
+  renderMain(true);``
 })();
