+++ index.css
@@ -5,6 +5,7 @@
 img {
   width: 24px;
   height: 24px;
+  cursor: pointer;
 }
 
 li {
